requirements.txt
colorama==0.4.6
crayons==0.4.0
numpy==2.2.6
opencv-python==4.11.0.86
phonenumbers==9.0.6
piexif==1.1.3
pillow==10.4.0
pycryptodome==3.23.0
qrcode==8.2
stegano==1.0.1
tqdm==4.67.1
